🩹 Allow big-object compilation for mingw gcc
MinGW GCC requires an option toggled for big-object compilation for the
assembler, otherwise it fails to compiler the unit tests due to the
Catch2 integration.

This adds the compiler flag when building for MinGW GCC.

This also removes building all of Catch2 so that it only builds the
relevant Catch2 components as part of the build.