🩹 Fix bad input to Delegate invocation
The `operator()` was incorrectly passing `m_storage` as the
argument for the storage, when in fact the value should be `this` as
it is passing the entire storage for the `delegate` type. This was
resulting in an incorrect cast to the wrong type.

This may not be "undefined behavior" on a technicality as the types are
trivial and standard-layout, and casting between the first element of
such a type to the owning type is valid in C++ -- but this was logically
incorrect.