🩹 Fix `clang-cl` -Wdocumentation warnings
`clang-cl` has a warning for `-Wdocumentation` which warns on incorrect
and invalid doxygen-style documentation for either misnamed parameters,
or for incorrect tags/attribution.

However this also displays a lot of false-positives. For example, on
the `delegate` class it detects `operator()`'s documentaton as being
incorrect due to `\return` being documented -- but this only flags if
you instantiate the template with a `void` return!

Additionally, it also complains that `\internal` is not a valid tag,
which it is: https://www.doxygen.nl/manual/commands.html#cmdinternal

These warnings have been addressed where possible, or suppressed in all
other cases.