🔨 Mark entire `union` as aligned
The buffer storage inside the union was previously the only marked
object for alignment with `max_align_t`. Although this works, this
also results in a non-uniform definition where it appears that only
the buffer has the alignment, but the other union members do not --
which is incorrect since all members of the union will have the same
alignment.

This reapplies the `alignas` specifier to instead be on the `union`
itself so that all union members are more clearly aligned to the same
boundary.