🔨 Change stub signature
The stub signature currently accepts a `void*` for the storage, which is
actually just a pointer to the `this` member of the delegate performing
the invocation.

It turns out that this cast to `void*` is actually unnecessary. On both
ends of the cast, the type is a `delegate*` -- which makes going through
`void*` an unnecessary conversion.

This has been removed.