🎯 Add tests for `delegate::bind`
The `delegate::bind` function was previously left untested since the
majority of the behavior simply delegates back to the `delegate`
constructor with a `bind` call. This was rather sloppy since it misses
a large number of cases where the SFINAE on the bind calls may be
incorrect.

Tests were added to validate this behavior.
As it turns out, the opaque function pointer `bind` calls were
implemented improperly previously -- and as such needed to be fixed
in the process.