🩹 Fix incorrect `constexpr` status
Several of the functions in `delegate` were marked `constexpr` despite
never being capable of yielding a core constant expression. functions
like `delegate::bind(...)` where the `...` refers to a constructor that
cannot, itself, be `constexpr` due to either the use of placement-`new`
or to the use of `reinterpret_cast`.

These have been fixed.

Additionally, stub functions that have incorrectly been marked as
_not_ `constexpr` despite being candidates for constant expressions have
also been marked `constexpr. These include only the statically bound
raw-function-pointer, and the empty-callables -- as these are the only
two cases that don't either require `reinterpret_cast` or casting
through `void*` -- neither of which are valid in a constant expression.

This now enables `operator()` to be properly `constexpr` for a select
few cases of bound functions.